BR-3875: Future PHP readiness 7.0 (#32220)
* BR-3875: Future PHP readiness 7.0

Version check updated

* BR-3805: Replace outdated Crypt_Blowfish library

Encryption replaced with OpenSSL, backward-compatible encryption is used

* BR-3875: Future PHP readiness 7.0

Arguments changed to be compatible with parent method, class property is being used instead

* BR-4314: Update Elastica library to make it PHP 7 compatible

* BR-3875: Future PHP readiness 7.0

Elasticsearch compatibility fixes

* BR-3875: Future PHP readiness 7.0

Arguments changed to be compatible with parent method, workaround for Home Activities

* BR-3875: Future PHP readiness 7.0

Fixes for methods signatures for PopupSmarty.
Fixes for methods signatures for DetailView - removed redundant inheritance, required utility methods copied to the child class.
Fixes for methods signatures for SearchForm2 and RelationshipHandler - removed redundant inheritance.

* BR-3875: Future PHP readiness 7.0

User-generated salt is deprecated for password_hash() in php7

* [BR-4207] Rework error handling in REST services

* BR-3875: Future PHP readiness 7.0

Don't use internal functions from polyfill library

* BR-3875: Future PHP readiness 7.0

Fixes for methods signature compatibility

* [BR-3875] Future PHP readiness 7.0

1. Array to string conversion notice
2. Only variables can be passed by reference

* BR-3875: Future PHP readiness 7.0

Static methods are declared and used as static

* BR-3875: Unit tests fixes

* BR-3875: Future PHP readiness 7.0

Calling mktime() and gmmktime() without arguments results in E_STRICT since php5.1 and E_DEPRECATED since php7. Change to use time() instead

* BR-3875: Future PHP readiness 7.0

Backward incompatible changes - methods signature fixes

* BR-3875: Future PHP readiness 7.0

Stable sort to keep keys ordering the same between different php versions

* BR-3875: Future PHP readiness 7.0 (#9)

Rename incompatible method

* BR-3875: Future PHP readiness 7.0

Methods signatures fixes

* BR-3875: Future PHP readiness 7.0

E_STRICT fixes

* BR-3875: Future PHP readiness 7.0

Fix notices

* BR-3875: Future PHP readiness 7.0

Compatibility fixes in tinyMCE

* Revert "Merge pull request #30979 from morozov/BR-4228"

This reverts commit 875df6273093fb9662f402b7c277045355b5e2f7, reversing
changes made to b1822e81aa7f05d994f538b05f4f140d3392775d.

* BR-3875: Future PHP readiness 7.0

onelogin/php-saml updated to 2.7.0 to fix php7 error - https://github.com/onelogin/php-saml/issues/98

* BR-3875 fix code style tests

* BR-3875 Fix php warnings and errors in unit tests

* BR-3875 fix unit test part 2

* [BR-3875] Reworked code for compatibitity with PHP 7.1

* [BR-3875] Got rid of undefined JSON_LOOSE_TYPE which is not ignored starting PHP 7.1

See https://bugs.php.net/bug.php?id=67829

* [BR-3875] Removed % sign from column width in list view definitions

* [BR-3875] Rework session storage to be compatible with PHP 7.1

Instead of suppressing warnings when reopening previously closed session, track the session ID manually and prevent sending HTTP headers.

* [BR-3875] Do not call session_regenerate_id() after destroying session, since it's supposed to be called when the session is active.

* BR-3875: Future PHP readiness 7.0

Another round of signature mismatch fixes from etsy/phan analyzer

* BR-3875: PHP 7.1 compatibility

Method compatibility for 7.1

* BR-3875: Future PHP readiness 7.0

Fix unit test

* [BR-3875] Removed some more PHP4-style constructors

+++ sugarcrm/tests/{old}/modules/DynamicFields/Bug61859_p2Test.php
@@ -90,7 +90,9 @@ class Bug61859_p2Test extends Sugar_PHPUnit_Framework_TestCase
 
     static public function tearDownAfterClass()
     {
-        self::$field->delete(self::$dynamicField);
+        if (isset(self::$field)) {
+            self::$field->delete(self::$dynamicField);
+        }
 
         SugarTestHelper::tearDown();
         parent::tearDownAfterClass();
